relax region id to int, too
we only allow positive numbers, like for the attribute id
we can encode none as 0 and end as negative id, without loosing any bit of encoding space
public api wraps that

+++ src/lib/ansihighlighter.cpp
@@ -1168,7 +1168,7 @@ private:
         int depth;
         int offset;
         int bindIndex;
-        quint16 regionId;
+        int regionId;
         State state;
     };
